moved more constants to the config class
