+++ app/constants/constants.py
@@ -34,8 +34,6 @@ INTRINSIC_RELATIONSHIPS = [
     "none",
 ]
 
-CONVERSATION_MESSAGE_RETENTION_COUNT = 10
-
 RESPOND_CHOICE = "respond"
 IGNORE_CHOICE = "ignore"
 USER_ROLE = "user"
@@ -991,11 +989,7 @@ MESSAGE_COLLECTION = "message"
 
 MEMORY_COLLECTION = "memory"
 
-EMOTIONAL_DECAY_RATE = 240
-
-MESSAGE_HISTORY_COUNT = 10
-
-THINKING_RATE = 600
+THOUGHT_COLLECTION = "thought"
 
 MYERS_BRIGGS_PERSONALITIES = [
   {
